% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_alert}
\alias{bs_alert}
\title{Create a Bootstrap alert}
\usage{
bs_alert(x, div_attr = c(class = "alert alert-primary"))
}
\arguments{
\item{x}{A string, the HTML or text to display in the alert.}

\item{div_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the div containing the alert.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/components/alerts/}.
}
\examples{
bs_alert(
"Hello"
)
}
