% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_badge}
\alias{bs_badge}
\title{Create a Bootstrap badge}
\usage{
bs_badge(x, span_attr = c(class = "badge bg-secondary"))
}
\arguments{
\item{x}{A string, the text to display in the badge (uses the <span> tag).}

\item{span_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the span containing the badge text.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/components/badge/}.
}
\examples{
bs_badge(
"Hello"
)
}
