% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_card}
\alias{bs_card}
\title{Create a Bootstrap card}
\usage{
bs_card(
  header = NULL,
  title = NULL,
  text = NULL,
  body,
  img_src = NULL,
  img_alt = NULL,
  div_attr = c(class = "card"),
  header_attr = c(class = "card-header"),
  title_attr = c(class = "card-title"),
  text_attr = c(class = "card-text"),
  body_attr = c(class = "card-body"),
  img_attr = c(class = "card-img-top")
)
}
\arguments{
\item{header}{A string, the HTML to display in the header of the card.}

\item{title}{A string, the HTML to display in the title of the card.}

\item{text}{A string, the HTML to display in the text of the card.}

\item{body}{A string, the HTML to display in the body of the card.}

\item{img_src}{A string, the path of an image to display with the card. Passed to "src" attribute of the <img> tag.}

\item{img_alt}{A string, the alt attribute of an image to display with the card. Passed to "alt" attribute of the <img> tag.}

\item{div_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the div wrapping the card content.}

\item{header_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the card header.}

\item{title_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the card title.}

\item{text_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the card text.}

\item{body_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the div wrapping the card body.}

\item{img_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the img tag if an image is to be displayed.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/components/card/}.
}
\examples{
bs_card(
body = "This is a card"
)
}
