% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_carousel}
\alias{bs_carousel}
\title{Create a Bootstrap carousel}
\usage{
bs_carousel(
  id,
  items = list(),
  carousel_class = "carousel slide",
  inner_class = "carousel-inner",
  item_class = "carousel-item",
  controls = TRUE
)
}
\arguments{
\item{id}{A string, the id to use for the carousel div.}

\item{items}{A list, entries should be HTML to display in the carousel.}

\item{carousel_class}{A string, the class of the div wrapping the carousel content.}

\item{inner_class}{A string, the class of the div wrapping the carousel items.}

\item{item_class}{A string, the class of the div wrapping each carousel item.}

\item{controls}{TRUE/FALSE, if TRUE, adds code to display arrows to click through the carousel.}
}
\value{
A string of HTML.
}
\description{
Create a Bootstrap carousel
}
\examples{
bs_carousel(
id = "c1",
items = list(
  h1("First slide"),
  h1("Second slide"),
  h1("Third Slide")
)
)
}
