% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_clearfix}
\alias{bs_clearfix}
\title{Create a div with class "clearfix"}
\usage{
bs_clearfix(...)
}
\arguments{
\item{...}{Valid string(s) of HTML.}
}
\value{
A string of HTML.
}
\description{
Create a div with class "clearfix"
}
\examples{
bs_clearfix(
p("Test")
)
}
