% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_col}
\alias{bs_col}
\title{Create a Bootstrap col div}
\usage{
bs_col(
  ...,
  size = NULL,
  align_self = NULL,
  order = NULL,
  offset = NULL,
  margin = NULL,
  custom_class = NULL
)
}
\arguments{
\item{...}{A string or strings of HTML content to include in the col div.}

\item{size}{An integer, 1-12, determines how many column units a particular column takes up.}

\item{align_self}{A string, likely one of "start", "center", or "end", that gets appended to "align-self-".}

\item{order}{An integer, sets the order of the column in a row.}

\item{offset}{A string, an offset value (ex. "md-3") to append to the "offset-" class.}

\item{margin}{A string, a valid Bootstrap margin class (ex. "me-auto")}

\item{custom_class}{A string, allows custom classes to be appended to the col class}
}
\value{
A string of HTML.
}
\description{
Create a Bootstrap col div
}
\examples{
bs_col(p("Col 1"))
}
