% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_col_break}
\alias{bs_col_break}
\title{Create a div with class "w-100" to force columns onto a new line}
\usage{
bs_col_break()
}
\value{
A string of HTML.
}
\description{
Create a div with class "w-100" to force columns onto a new line
}
\examples{
bs_col_break()
}
