% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_collapse}
\alias{bs_collapse}
\title{Create a Bootstrap collapse}
\usage{
bs_collapse(
  id,
  button_label,
  collapse_content,
  button_class = "btn btn-primary",
  div_class = "collapse"
)
}
\arguments{
\item{id}{A string, the id to use for the collapse div.}

\item{button_label}{A string, the text to display in the button controlling the collapsible content.}

\item{collapse_content}{A string, the HTML to display or collapse.}

\item{button_class}{A string, the class added to the button controlling the collapse.}

\item{div_class}{A string, the class added to the div wrapping the collapsible content.}
}
\value{
A string of HTML.
}
\description{
Create a Bootstrap collapse
}
\examples{
bs_collapse(
id = "collapse1",
button_label = "Click to Expand",
collapse_content = p("Hello")
)
}
