% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_container}
\alias{bs_container}
\title{Create a Bootstrap container div}
\usage{
bs_container(
  ...,
  fluid = FALSE,
  breakpoint = NULL,
  padding = NULL,
  overflow_hidden = FALSE,
  custom_class = NULL
)
}
\arguments{
\item{...}{A string or strings of HTML content to include in the container div.}

\item{fluid}{TRUE/FALSE, if TRUE, the class becomes "container-fluid" which sets the width to 100\%.}

\item{breakpoint}{A string, sets the break-point for reshaping content based on screen size. One of "sm", "md", "lg", "xl", "xxl".}

\item{padding}{An integer, the padding to add, measured in px.}

\item{overflow_hidden}{TRUE/FALSE, if TRUE, adds "overflow-hidden" to the container class}

\item{custom_class}{A string, allows custom classes to be appended to the container class}
}
\value{
A string of HTML.
}
\description{
Create a Bootstrap container div.
}
\examples{
bs_container(
div(),
fluid = TRUE
)
}
