% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_doc}
\alias{bs_doc}
\title{Create a HTML document preconfigured to load Bootstrap5 from the CDN}
\usage{
bs_doc(head = NULL, body)
}
\arguments{
\item{head}{A string of HTML that will be passed to the html5::head() function.}

\item{body}{A string of HTML that will be passed to the html5::body() function}
}
\value{
A string of HTML.
}
\description{
Create a HTML document preconfigured to load Bootstrap5 from the CDN
}
\examples{
bs_doc(
body = div(h1("Hello"), p("Welcome to this page."))
)
}
