% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_figure}
\alias{bs_figure}
\title{Create HTML figure tag configured with Bootstrap classes}
\usage{
bs_figure(
  src,
  alt = NULL,
  caption = NULL,
  figure_class = "figure",
  img_class = "figure-img img-fluid",
  figcaption_class = "figure-caption"
)
}
\arguments{
\item{src}{A string, the value to pass to the "src" param of html5::img().}

\item{alt}{A string, the value to pass to the "alt" html5::img().}

\item{caption}{A string, gets passed to html5::figcaption().}

\item{figure_class}{A string, the class of the figure tag.}

\item{img_class}{A string, the class of the img tag.}

\item{figcaption_class}{A string, the class of the figcaption tag.}
}
\value{
A string of HTML.
}
\description{
Create HTML figure tag configured with Bootstrap classes
}
\examples{
bs_figure(
src = tempfile(),
caption = "This is an example"
)
}
