% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_flexbox}
\alias{bs_flexbox}
\title{Create a Bootstrap flexbox}
\usage{
bs_flexbox(x, div_attr = c(class = "d-flex justify-content-center"))
}
\arguments{
\item{x}{A string, the HTML or text to display in the div.}

\item{div_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the div containing x. Defaults to center the content of x.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/utilities/flex/}.
}
\examples{
bs_flexbox(
p("Hello")
)
}
