% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_row}
\alias{bs_row}
\title{Create a Bootstrap row div}
\usage{
bs_row(
  ...,
  row_cols = NULL,
  align_items = NULL,
  justify_content = NULL,
  horizontal_gutter = NULL,
  vertical_gutter = NULL,
  gutter = NULL,
  custom_class = NULL
)
}
\arguments{
\item{...}{A string or strings of HTML content to include in the row div.}

\item{row_cols}{An integer, the number of columns to use for rendering content.}

\item{align_items}{A string, likely one of "start", "center", or "end", that gets appended to "align-items-".}

\item{justify_content}{A string, such as "md-center", that gets appended to "justify-content-".}

\item{horizontal_gutter}{An integer, controls the horizontal gutter width, gets appended to "gx-".}

\item{vertical_gutter}{An integer, controls the vertical gutter width, gets appended to "gy-".}

\item{gutter}{An integer, controls the horizontal and vertical gutter width, gets appended to "g-".}

\item{custom_class}{A string, allows custom classes to be appended to the row class}
}
\value{
A string of HTML.
}
\description{
Create a Bootstrap row div
}
\examples{
bs_row(
bs_col(p("Col 1")),
bs_col(p("Col 2"))
)
}
