% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_spinner}
\alias{bs_spinner}
\title{Create a Bootstrap spinner}
\usage{
bs_spinner(div_attr = c(class = "spinner-border text-primary"))
}
\arguments{
\item{div_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the div displaying the spinner.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/components/spinners/}.
}
\examples{
bs_spinner()
}
