% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_toast_js}
\alias{bs_toast_js}
\title{Create a <script> tag to initialize the Bootstrap javascript for creating toasts. Must include in HTML document if using a toast.}
\usage{
bs_toast_js()
}
\value{
A string of HTML.
}
\description{
Create a <script> tag to initialize the Bootstrap javascript for creating toasts. Must include in HTML document if using a toast.
}
\examples{
bs_toast_js()
}
