% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\docType{data}
\name{css_href_var}
\alias{css_href_var}
\title{Variable for the CSS CDN link}
\format{
A string
}
\usage{
css_href_var
}
\description{
"https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
}
\keyword{datasets}
