% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\docType{data}
\name{css_integrity_var}
\alias{css_integrity_var}
\title{Variable for the CSS CDN integrity string}
\format{
A string
}
\usage{
css_integrity_var
}
\description{
"sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
}
\keyword{datasets}
