% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\docType{data}
\name{js_integrity_var}
\alias{js_integrity_var}
\title{Variable for the JS CDN integrity string}
\format{
A string
}
\usage{
js_integrity_var
}
\description{
"sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
}
\keyword{datasets}
