% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\docType{data}
\name{js_src_var}
\alias{js_src_var}
\title{Variable for the JS CDN link}
\format{
A string
}
\usage{
js_src_var
}
\description{
"https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
}
\keyword{datasets}
