% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs_card}
\alias{bs_card}
\title{Create a Bootstrap card}
\usage{
bs_card(
  body,
  img_src = NULL,
  img_alt = NULL,
  div_class = "card",
  body_class = "card-body",
  img_class = "card-img-top"
)
}
\arguments{
\item{body}{A string, the HTML to display in the body of the card.}

\item{img_src}{A string, the path of an image to display with the card. Passed to "src" param of html5::img().}

\item{img_alt}{A string, the alt attribute of an image to display with the card. Passed to "alt" param of html5::img().}

\item{div_class}{A string, the class of the div wrapping the card content.}

\item{body_class}{A string, the class of the div wrapping the card body.}

\item{img_class}{A string, the class of the img tag if an image is to be displayed.}
}
\value{
A string of HTML.
}
\description{
Create a Bootstrap card
}
\examples{
bs_card(
body = "This is a card"
)
}
