% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_clearfix}
\alias{bs_clearfix}
\title{Create a div with class "clearfix"}
\usage{
bs_clearfix(..., clearfix_attr = c(class = "clearfix"))
}
\arguments{
\item{...}{Valid string(s) of HTML.}

\item{clearfix_attr}{A named list or named vector, names are attribute names and values are attribute values.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/layout/columns/}.
}
\examples{
bs_clearfix(
p("Test")
)
}
