% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_col}
\alias{bs_col}
\title{Create a Bootstrap col div}
\usage{
bs_col(..., col_attr = c(class = "col"))
}
\arguments{
\item{...}{A string or strings of HTML content to include in the col div.}

\item{col_attr}{A named list or named vector, names are attribute names and values are attribute values.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/layout/columns/}.
}
\examples{
bs_col(p("Col 1"))
}
