% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_col_break}
\alias{bs_col_break}
\title{Create a div with class "w-100" to force columns onto a new line}
\usage{
bs_col_break(col_break_attr = c(class = "w-100"))
}
\arguments{
\item{col_break_attr}{A named list or named vector, names are attribute names and values are attribute values.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/layout/columns/}.
}
\examples{
bs_col_break()
}
