% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_collapse}
\alias{bs_collapse}
\title{Create a Bootstrap collapse}
\usage{
bs_collapse(
  id,
  button_label,
  content,
  button_attr = c(class = "btn btn-primary"),
  content_attr = c(class = "collapse")
)
}
\arguments{
\item{id}{A string, the id to use for the collapse div.}

\item{button_label}{A string, the text to display in the button controlling the collapsible content.}

\item{content}{A string, the HTML to display or collapse.}

\item{button_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the button controlling the collapse.}

\item{content_attr}{A named list or named vector, names are attribute names and values are attribute values. Added to the div wrapping the collapsible content.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/components/collapse/}.
}
\examples{
bs_collapse(
id = "collapse1",
button_label = "Click to Expand",
content = p("Hello")
)
}
