% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_container}
\alias{bs_container}
\title{Create a Bootstrap container div}
\usage{
bs_container(..., container_attr = c(class = "container"))
}
\arguments{
\item{...}{A string or strings of HTML content to include in the container div.}

\item{container_attr}{A named list or named vector, names are attribute names and values are attribute values.}
}
\value{
A string of HTML.
}
\description{
Create a Bootstrap container div.
}
\details{
Learn more at \url{https://getbootstrap.com/docs/5.1/layout/containers/}.
}
\examples{
bs_container(
div()
)
}
