% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_doc}
\alias{bs_doc}
\title{Create a HTML document preconfigured to load Bootstrap5 from the CDN}
\usage{
bs_doc(
  head = NULL,
  body,
  body_attr = NULL,
  css_href = css_href_var,
  css_integrity = css_integrity_var,
  js_src = js_src_var,
  js_integrity = js_integrity_var
)
}
\arguments{
\item{head}{A string of HTML that will be passed to the ... param of the html5::head() function.}

\item{body}{A string of HTML that will be passed to the ... param of the html5::body() function.}

\item{body_attr}{A named list or named vector, passed to the attr param of the html5::body() function.}

\item{css_href}{The url of jsDelivr for the version of bootstrap css to include.}

\item{css_integrity}{The integrity string for the version of bootstrap css to include.}

\item{js_src}{The url of jsDelivr for the version of bootstrap js bundle to include.}

\item{js_integrity}{The integrity string for the version of bootstrap js bundle to include.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/getting-started/introduction/}.
}
\examples{
bs_doc(
body = div(h1("Hello"), p("Welcome to this page."))
)
}
