% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_figure}
\alias{bs_figure}
\title{Create HTML figure tag configured with Bootstrap classes}
\usage{
bs_figure(
  figure_attr = c(class = "figure"),
  img_attr = c(src = "...", alt = "...", class = "figure-img img-fluid"),
  figcaption_attr = c(class = "figure-caption"),
  figcaption = NULL
)
}
\arguments{
\item{figure_attr}{A named list or named vector, names are attribute names and values are attribute values. Passed to the attr parameter of html5::figure.}

\item{img_attr}{A named list or named vector, names are attribute names and values are attribute values. Passed to the attr parameter of html5::img.}

\item{figcaption_attr}{A named list or named vector, names are attribute names and values are attribute values. Passed to the attr parameter of html5::figcaption.}

\item{figcaption}{A string, gets passed to html5::figcaption.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/content/figures/}.
}
\examples{
bs_figure(
img_attr = c(
"src" = "...",
"alt" = "...",
"class" = "figure-img img-fluid"
)
)
}
