% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_img}
\alias{bs_img}
\title{Create HTML img tag configured with Bootstrap classes}
\usage{
bs_img(img_attr = c(src = "...", alt = "...", class = "img"), center = FALSE)
}
\arguments{
\item{img_attr}{A named list or named vector, names are attribute names and values are attribute values. Passed to the attr parameter of html5::img.}

\item{center}{TRUE/FALSE, if TRUE, wraps img tag in a div with class "text-center".}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/content/images/}.
}
\examples{
bs_img(
img_attr = c(
"src" = "...",
"alt" = "...",
"class" = "img"
)
)
}
