% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTools2.R
\name{bs_row}
\alias{bs_row}
\title{Create a Bootstrap row div}
\usage{
bs_row(..., row_attr = c(class = "row"))
}
\arguments{
\item{...}{A string or strings of HTML content to include in the row div.}

\item{row_attr}{A named list or named vector, names are attribute names and values are attribute values.}
}
\value{
A string of HTML.
}
\description{
Learn more at \url{https://getbootstrap.com/docs/5.1/layout/grid/}.
}
\examples{
bs_row(
bs_col(p("Col 1")),
bs_col(p("Col 2"))
)
}
